HBX-1674: Move class 'org.hibernate.tool.hbmlint.detector.TableSelectorStrategy' to 'org.hibernate.tool.internal.reveng.TableSelectorStrategy'
