+++ orm/src/main/java/org/hibernate/tool/internal/reveng/TableSelectorStrategy.java
@@ -1,14 +1,13 @@
 /**
  * 
  */
-package org.hibernate.tool.hbmlint.detector;
+package org.hibernate.tool.internal.reveng;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.hibernate.tool.api.reveng.ReverseEngineeringStrategy;
 import org.hibernate.tool.api.reveng.SchemaSelection;
-import org.hibernate.tool.internal.reveng.DelegatingReverseEngineeringStrategy;
 
 public class TableSelectorStrategy extends DelegatingReverseEngineeringStrategy {
